SupersonicBuilder/requirements.txt
# PDF Generation
reportlab
Pillow
pypdf
PyPDF2
PyYAML
qrcode[pil]
svglib
pdf2image
pikepdf

# Serial & Monitoring
pyserial
watchdog
psutil

# Web Framework
Flask>=3.0.0
flask-cors
fpdf
fpdf2

# Verify / Auto-fix
beautifulsoup4>=4.12.0
requests>=2.32.0

# Voice Commander
pyttsx3>=2.90

# Production Server
gunicorn>=21.2

# Supersonic Control Core v4 Dependencies
colorama>=0.4.6
rich>=13.0.0
# playsound==1.3.0  # Removed - installation issues
semver>=3.0.2

# Optional AI and signing integrations (v4)
openai>=1.0.0
langchain>=0.1.0

# Note: SBOM/security tools are external binaries
# Install separately: trivy, syft, cosign, adb (Android platform-tools)
flask
psutil
pytest
pytest-cov
